+++ StudentSoup/src/main/java/ProjectDoge/StudentSoup/exception/boardreview/BoardReviewContentNullException.java
@@ -1,4 +1,4 @@
-package ProjectDoge.StudentSoup.exception.board;
+package ProjectDoge.StudentSoup.exception.boardreview;
 
 public class BoardReviewContentNullException extends RuntimeException{
     public BoardReviewContentNullException() { super(); }
