+++ StudentSoup/src/main/java/ProjectDoge/StudentSoup/exception/boardreview/BoardReviewIdNotSentException.java
@@ -1,4 +1,4 @@
-package ProjectDoge.StudentSoup.exception.board;
+package ProjectDoge.StudentSoup.exception.boardreview;
 
 public class BoardReviewIdNotSentException extends RuntimeException{
     public BoardReviewIdNotSentException() {
