+++ StudentSoup/src/main/java/ProjectDoge/StudentSoup/exception/boardreview/BoardReviewNotOwnException.java
@@ -1,4 +1,4 @@
-package ProjectDoge.StudentSoup.exception.board;
+package ProjectDoge.StudentSoup.exception.boardreview;
 
 public class BoardReviewNotOwnException extends RuntimeException{
