+++ StudentSoup/src/main/java/ProjectDoge/StudentSoup/service/BoardReview/BoardReviewRegisterService.java
@@ -5,7 +5,7 @@ import ProjectDoge.StudentSoup.dto.board.BoardReviewResDto;
 import ProjectDoge.StudentSoup.entity.board.Board;
 import ProjectDoge.StudentSoup.entity.board.BoardReview;
 import ProjectDoge.StudentSoup.entity.member.Member;
-import ProjectDoge.StudentSoup.exception.board.BoardReviewContentNullException;
+import ProjectDoge.StudentSoup.exception.boardreview.BoardReviewContentNullException;
 import ProjectDoge.StudentSoup.repository.boardreview.BoardReviewRepository;
 import ProjectDoge.StudentSoup.service.board.BoardFindService;
 import ProjectDoge.StudentSoup.service.member.MemberFindService;
@@ -29,14 +29,14 @@ public class BoardReviewRegisterService {
     @Transactional
     public Long join(BoardReviewResDto dto){
         log.info("게시글 리뷰 등록 서비스가 실행됐습니다.");
-        checkNullContent(dto);
+        checkContent(dto);
         BoardReview boardReview = createBoardReview(dto);
         BoardReview review = boardReviewRepository.save(boardReview);
         log.info("게시글 리뷰 등록 서비스가 실행됐습니다.");
         return review.getReviewId();
     }
 
-    private void checkNullContent(BoardReviewResDto dto) {
+    private void checkContent(BoardReviewResDto dto) {
         if(dto.getContent().strip().length()==0 || dto.getContent().length()==0){
             throw new BoardReviewContentNullException("리뷰 내용이 null 또는 빈칸 입니다.");
         }
